+++ src/Makefile.am
@@ -54,7 +54,7 @@ noinst_HEADERS = \
 	conf-yaml-loader.h \
 	conf.h \
 	counters.h \
-	datajson.h \
+	datasets-context-json.h \
 	datasets-ipv4.h \
 	datasets-ipv6.h \
 	datasets-md5.h \
@@ -653,7 +653,7 @@ libsuricata_c_a_SOURCES = \
 	conf-yaml-loader.c \
 	conf.c \
 	counters.c \
-	datajson.c \
+	datasets-context-json.c \
 	datasets-ipv4.c \
 	datasets-ipv6.c \
 	datasets-md5.c \
