+++ src/runmode-unix-socket.c
@@ -55,7 +55,7 @@
 #include "conf-yaml-loader.h"
 
 #include "datasets.h"
-#include "datajson.h"
+#include "datasets-context-json.h"
 #include "runmode-unix-socket.h"
 
 int unix_socket_mode_is_running = 0;
